Reduce copying for 20x faster performance
Changed parts of the loading process so that the numpy array from
librosa.load is stored directly into the sources array.
Also, duration is now taken into account in the load() function
instead of manually trimming in a separate line.

On a duration of 1 second with 4 mp3s, time was reduced from 5 seconds
down to 262 milliseconds, around 20 times faster.