Complete cleaning script
The cleaning script is now complete. It scans the directory and attempts
to load info on each song. The process is now very fast due to using
the torchaudio.info() rather than torchaudio.load() to check for
corruption.

Corrupted files are now listed. The user will be asked to confirm
deleting the selected files.

Lastly, some of the files seem to show a strange warning:
MAD lost sync
These are not throwing exceptions, so for now I will assume that
there is no trouble with these files.